refactored algorithms
