+++ doc/examples/nn_model_summary.py
@@ -1,5 +1,5 @@
-import light_prop.propagation.methods as prop
-from light_prop.propagation.params import PropagationParams
+import lightprop.propagation.methods as prop
+from lightprop.propagation.params import PropagationParams
 
 params = PropagationParams.get_example_propagation_data()
 propagator = prop.NNPropagation(params)
