+++ lightprop/algorithms/nn_trainer.py
@@ -1,37 +1,9 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Aug  4 14:09:21 2022
-
-@author: dr PK & be BB with mse MD
-"""
-
 import matplotlib.pyplot as plt
 import tensorflow as tf
 from tensorflow import keras
-
 import lightprop.propagation.methods as prop
 from lightprop.lightfield import LightField
 from lightprop.propagation.params import PropagationParams
-
-
-class GerchbergSaxton:
-    def __init__(self, propagation_params: PropagationParams):
-        self.params = propagation_params
-
-    def optimize(self, input_field: LightField, target_field: LightField, iterations: int = 5):
-        output_plane = target_field
-        input_plane = None
-
-        for i in range(iterations):
-            output_plane.amplitude = target_field.amplitude
-            self.params.distance *= -1
-            input_plane = prop.ConvolutionPropagation(self.params).propagate(output_plane)
-            input_plane.amplitude = input_field.amplitude
-            self.params.distance *= -1
-            output_plane = prop.ConvolutionPropagation(self.params).propagate(input_plane)
-
-        return (input_plane, output_plane)
-
 
 class NNTrainer:
     def __init__(self, propagation_params: PropagationParams):
